fix: low gauss to index

+++ utils/draw.py
@@ -52,8 +52,8 @@ def draw_detector(background, detector: Line):
     for i in range(foreground.shape[0]):
         for j in range(foreground.shape[1]):
             for col in range(3):
-                x = detector.x1 + i / foreground.shape[0] * detector.x2
-                y = detector.y1 + j / foreground.shape[1] * detector.y2
+                x = int(detector.x1 + i / foreground.shape[0] * detector.x2)
+                y = int(detector.y1 + j / foreground.shape[1] * detector.y2)
                 background[x, y, col] = alpha_foreground[i, j] * foreground[i, j, col] + \
                     alpha_background[x, y] * background[x, y, col] * (1 - alpha_foreground[i, j])
